Closes #8

+++ db_provision.sql
@@ -30,8 +30,8 @@ CREATE UNIQUE INDEX t_pkey ON tokens("yahooTokenRetrievedTime" int8_ops);
 
 CREATE TABLE messages (
     name text,
-    message_id integer PRIMARY KEY,
-    user_id integer,
+    message_id bigint PRIMARY KEY,
+    user_id bigint,
     message text,
     time bigint
 );
